requirements.txt
numpy
Pillow==5.1.0
scikit-learn==0.19.1
torch==0.4.0
opencv-python==3.4.0.12
visdom==0.1.7
torchvision=0.2.1